requirements.txt
adjustText==0.8
altair==4.2.2
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==22.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.2.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
eli5==0.13.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.1
h11==0.14.0
highlight-text==0.2
httpcore==1.0.7
httpx==0.28.1
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.4
lightgbm==3.3.5
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mplsoccer==1.1.10
narwhals==1.26.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.24.2
openpyxl==3.1.1
overrides==7.7.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.6
plotly==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==11.0.0
pycparser==2.22
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==3.2.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rpds-py==0.22.3
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.19.0
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.1.0
tinycss2==1.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.2
uri-template==1.3.0
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xgboost==1.7.4
zipp==3.15.0
pyyaml
